Some basic Hit detection attempt.
At the moment, it detects when the player sprite hits the top, left and
right of a test GameObject called block. For some reason, I can't get it
to detect when the player collides with the bottom of the block and when
the player collides with the sides, it seems to "teleport" it to a set
position above the block. No idea why.